+++ client/src/components/Auth.jsx
@@ -4,6 +4,43 @@ import { Link } from 'react-router-dom'
 import { useState, useEffect } from 'react'
 import { requestAuth } from '../utils/AuthUtil'
 
+const useCurrentLocation = (options = {}) => {
+    // location 정보 저장
+    const [location, setLocation] = useState()
+    // 에러 메세지 저장
+    const [error, setError] = useState()
+  
+    // Geolocation의 `getCurrentPosition` 메소드에 대한 성공 callback 핸들러
+    const handleSuccess = (pos) => {
+        const { latitude, longitude } = pos.coords
+    
+        setLocation({
+            latitude,
+            longitude,
+        })
+    };
+  
+    // Geolocation의 `getCurrentPosition` 메소드에 대한 실패 callback 핸들러
+    const handleError = (error) => {
+        setError(error.message)
+    };
+  
+    useEffect(() => {
+        const { geolocation } = navigator
+    
+        // 사용된 브라우저에서 지리적 위치(Geolocation)가 정의되지 않은 경우 오류로 처리합니다.
+        if (!geolocation) {
+            setError('Geolocation is not supported.')
+            return
+        }
+  
+        // Geolocation API 호출
+        geolocation.getCurrentPosition(handleSuccess, handleError, options)
+    }, [options])
+
+    return { location, error }
+}
+
 const Auth = ({mode}) => {
     const dispatch = useDispatch()
     const navigate = useNavigate()
@@ -11,10 +48,16 @@ const Auth = ({mode}) => {
         {
             name: '',
             email: '',
-            password: ''
+            password: '',
+            location: {latitude: -1, longitude: -1}
         }
     )
     const [err, setErr] = useState(false)
+
+    // check gps location 
+    const { location, geoError } = useCurrentLocation()
+
+    console.log(location, geoError)
 
     useEffect(() => {
         requestAuth(mode, 'get', '', dispatch, navigate)
@@ -22,7 +65,7 @@ const Auth = ({mode}) => {
     }, [mode])
 
     const handleChange = (event) => {
-        setValue({...value, [event.target.id]: event.target.value})    
+        setValue({...value, [event.target.id]: event.target.value, location})    
     }
 
     const handleSubmit = (event) => {