+++ client/src/components/Sidebar.jsx
@@ -39,6 +39,9 @@ const Logo = styled.span`
     font-size: 20px;
     font-weight: bold;
     color: #6439ff;
+    @media screen and (max-width: 500px) {
+        font-size: 12px;
+    }
 `
 
 const Items = styled.ul`
@@ -68,6 +71,10 @@ const Item = styled.li`
 
     border: ${props => props.active ? '1px solid black' : 'none'};
     border-radius: ${props => props.active ? '20px' : 'none'};
+
+    @media screen and (max-width: 500px) {
+        height: 30px;
+    }
 ` 
 
 const Icon = styled.div`
@@ -80,6 +87,9 @@ const Span = styled.span`
     font-weight: 600;
     color: #888;
     margin-left: 10px;
+    @media screen and (max-width: 500px) {
+        display: none;
+    }
 `
 
 const Color = styled.div`